Split `stdlib.ts` into a `MATH` and `MISC` group
