+++ src/stdlib/math.ts
@@ -1,333 +1,23 @@
-import {
-  BigIntValue,
-  BoolValue,
-  BuiltinValue,
-  ComplexValue,
-  NumberValue,
-  StringValue,
-  Value,
-} from "./engines/cse/stash";
-// npm install mathjs
 import { erf, gamma, lgamma } from "mathjs";
-import { Context } from "./engines/cse/context";
-import { ControlItem } from "./engines/cse/control";
-import { handleRuntimeError } from "./engines/cse/error";
-import { displayOutput, receiveInput } from "./engines/cse/streams";
-import {
-  MissingRequiredPositionalError,
-  TooManyPositionalArgumentsError,
-  TypeError,
-  UserError,
-  ValueError,
-} from "./errors/errors";
-import { stringify } from "./utils/stringify";
+import { ExprNS } from "../ast-types";
+import { Context } from "../engines/cse/context";
+import { handleRuntimeError } from "../engines/cse/error";
+import { BigIntValue, BoolValue, BuiltinValue, NumberValue, Value } from "../engines/cse/stash";
+import { isNumeric } from "../engines/cse/utils";
+import { TypeError, ValueError } from "../errors";
+import { GroupName, minArgMap, Validate } from "./utils";
 
-export const minArgMap = new Map<string, number>();
+const mathBuiltins = new Map<string, Value>();
 
-export function Validate<T extends Value | Promise<Value>>(
-  minArgs: number | null,
-  maxArgs: number | null,
-  functionName: string,
-  strict: boolean,
-) {
-  return function (
-    _target: unknown,
-    _propertyKey: string,
-    descriptor: TypedPropertyDescriptor<
-      (args: Value[], source: string, command: ExprNS.Call, context: Context) => T
-    >,
-  ): void {
-    const originalMethod = descriptor.value!;
-    minArgMap.set(functionName, minArgs || 0);
-    descriptor.value = function (
-      args: Value[],
-      source: string,
-      command: ExprNS.Call,
-      context: Context,
-    ): T {
-      if (minArgs !== null && args.length < minArgs) {
-        handleRuntimeError(
-          context,
-          new MissingRequiredPositionalError(source, command, functionName, minArgs, args, strict),
-        );
-      }
+const constantMap = {
+  math_e: { type: "number", value: Math.E },
+  math_inf: { type: "number", value: Infinity },
+  math_nan: { type: "number", value: NaN },
+  math_pi: { type: "number", value: Math.PI },
+  math_tau: { type: "number", value: 2 * Math.PI },
+} as const;
 
-      if (maxArgs !== null && args.length > maxArgs) {
-        handleRuntimeError(
-          context,
-          new TooManyPositionalArgumentsError(source, command, functionName, maxArgs, args, strict),
-        );
-      }
-
-      return originalMethod.call(this, args, source, command, context);
-    };
-  };
-}
-
-export class BuiltInFunctions {
-  @Validate(1, 1, "arity", true)
-  static arity(args: Value[], source: string, command: ExprNS.Call, context: Context): BigIntValue {
-    const func = args[0];
-    if (func.type !== "builtin" && func.type !== "closure") {
-      handleRuntimeError(context, new TypeError(source, command, context, func.type, "function"));
-    }
-    if (func.type === "closure") {
-      const variadicInstance = func.closure.node.parameters.findIndex(param => param.isStarred);
-      if (variadicInstance !== -1) {
-        return { type: "bigint", value: BigInt(variadicInstance) };
-      }
-      return { type: "bigint", value: BigInt(func.closure.node.parameters.length) };
-    }
-    return { type: "bigint", value: BigInt(func.minArgs) };
-  }
-
-  @Validate(null, 2, "int", true)
-  static int(args: Value[], source: string, command: ExprNS.Call, context: Context): BigIntValue {
-    if (args.length === 0) {
-      return { type: "bigint", value: BigInt(0) };
-    }
-    const arg = args[0];
-    if (!isNumeric(arg) && arg.type !== "string" && arg.type !== "bool") {
-      handleRuntimeError(
-        context,
-        new TypeError(source, command, context, arg.type, "str, int, float or bool"),
-      );
-    }
-
-    if (args.length === 1) {
-      if (arg.type === "number") {
-        const truncated = Math.trunc(arg.value);
-        return { type: "bigint", value: BigInt(truncated) };
-      }
-      if (arg.type === "bigint") {
-        return { type: "bigint", value: arg.value };
-      }
-      if (arg.type === "string") {
-        const str = arg.value.trim().replace(/_/g, "");
-        if (!/^[+-]?\d+$/.test(str)) {
-          handleRuntimeError(context, new ValueError(source, command, context, "int"));
-        }
-        return { type: "bigint", value: BigInt(str) };
-      }
-      return { type: "bigint", value: arg.value ? BigInt(1) : BigInt(0) };
-    }
-    const baseArg = args[1];
-    if (arg.type !== "string") {
-      handleRuntimeError(context, new TypeError(source, command, context, arg.type, "string"));
-    }
-    if (baseArg.type !== "bigint") {
-      handleRuntimeError(context, new TypeError(source, command, context, baseArg.type, "int"));
-    }
-
-    let base = Number(baseArg.value);
-    let str = arg.value.trim().replace(/_/g, "");
-
-    const sign = str.startsWith("-") ? -1 : 1;
-    if (str.startsWith("+") || str.startsWith("-")) {
-      str = str.substring(1);
-    }
-
-    if (base === 0) {
-      if (str.startsWith("0x") || str.startsWith("0X")) {
-        base = 16;
-        str = str.substring(2);
-      } else if (str.startsWith("0o") || str.startsWith("0O")) {
-        base = 8;
-        str = str.substring(2);
-      } else if (str.startsWith("0b") || str.startsWith("0B")) {
-        base = 2;
-        str = str.substring(2);
-      } else {
-        base = 10;
-      }
-    }
-
-    if (base < 2 || base > 36) {
-      handleRuntimeError(context, new ValueError(source, command, context, "int"));
-    }
-
-    const validChars = "0123456789abcdefghijklmnopqrstuvwxyz".substring(0, base);
-    const regex = new RegExp(`^[${validChars}]+$`, "i");
-    if (!regex.test(str)) {
-      handleRuntimeError(context, new ValueError(source, command, context, "int"));
-    }
-
-    let res = BigInt(0);
-    for (const char of str) {
-      res = res * BigInt(base) + BigInt(validChars.indexOf(char.toLowerCase()));
-    }
-    return { type: "bigint", value: BigInt(sign) * res };
-  }
-
-  @Validate(null, 1, "float", true)
-  static float(args: Value[], source: string, command: ExprNS.Call, context: Context): NumberValue {
-    if (args.length === 0) {
-      return { type: "number", value: 0 };
-    }
-    const val = args[0];
-    if (val.type === "bigint") {
-      return { type: "number", value: Number(val.value) };
-    } else if (val.type === "number") {
-      return { type: "number", value: val.value };
-    } else if (val.type === "bool") {
-      return { type: "number", value: val.value ? 1 : 0 };
-    } else if (val.type === "string") {
-      const str = val.value.trim().replace(/_/g, "").toLowerCase();
-      const mappings = {
-        inf: Infinity,
-        "+inf": Infinity,
-        "-inf": -Infinity,
-        infinity: Infinity,
-        "+infinity": Infinity,
-        "-infinity": -Infinity,
-        nan: NaN,
-        "+nan": NaN,
-        "-nan": NaN,
-      };
-      if (str in mappings) {
-        return { type: "number", value: mappings[str as keyof typeof mappings] };
-      }
-      const num = Number(str);
-      if (isNaN(num)) {
-        handleRuntimeError(context, new ValueError(source, command, context, "float"));
-      }
-      return { type: "number", value: num };
-    }
-    handleRuntimeError(
-      context,
-      new TypeError(source, command, context, val.type, "'float', 'int', 'bool' or 'str'"),
-    );
-  }
-
-  @Validate(null, 2, "complex", true)
-  static complex(
-    args: Value[],
-    source: string,
-    command: ExprNS.Call,
-    context: Context,
-  ): ComplexValue {
-    if (args.length === 0) {
-      return { type: "complex", value: new PyComplexNumber(0, 0) };
-    }
-    if (args.length == 1) {
-      const val = args[0];
-      if (
-        val.type !== "bigint" &&
-        val.type !== "number" &&
-        val.type !== "bool" &&
-        val.type !== "string" &&
-        val.type !== "complex"
-      ) {
-        handleRuntimeError(context, new TypeError(source, command, context, val.type, "complex"));
-      }
-      return {
-        type: "complex",
-        value: PyComplexNumber.fromValue(context, source, command, val.value),
-      };
-    }
-    const invalidType = args.filter(
-      val =>
-        val.type !== "bigint" &&
-        val.type !== "number" &&
-        val.type !== "bool" &&
-        val.type !== "complex",
-    );
-    if (invalidType.length > 0) {
-      handleRuntimeError(
-        context,
-        new TypeError(
-          source,
-          command,
-          context,
-          invalidType[0].type,
-          "'int', 'float', 'bool' or 'complex'",
-        ),
-      );
-    }
-    const [real, imag] = args as (BigIntValue | NumberValue | BoolValue | ComplexValue)[];
-    const realPart = PyComplexNumber.fromValue(context, source, command, real.value);
-    const imagPart = PyComplexNumber.fromValue(context, source, command, imag.value);
-    return { type: "complex", value: realPart.add(imagPart.mul(new PyComplexNumber(0, 1))) };
-  }
-
-  @Validate(1, 1, "real", true)
-  static real(args: Value[], source: string, command: ExprNS.Call, context: Context): NumberValue {
-    const val = args[0];
-    if (val.type !== "complex") {
-      handleRuntimeError(context, new TypeError(source, command, context, val.type, "complex"));
-    }
-    return { type: "number", value: val.value.real };
-  }
-
-  @Validate(1, 1, "imag", true)
-  static imag(args: Value[], source: string, command: ExprNS.Call, context: Context): NumberValue {
-    const val = args[0];
-    if (val.type !== "complex") {
-      handleRuntimeError(context, new TypeError(source, command, context, val.type, "complex"));
-    }
-    return { type: "number", value: val.value.imag };
-  }
-
-  @Validate(null, 1, "bool", true)
-  static bool(args: Value[], _source: string, _command: ControlItem, _context: Context): BoolValue {
-    if (args.length === 0) {
-      return { type: "bool", value: false };
-    }
-    const val = args[0];
-    return { type: "bool", value: !isFalsy(val) };
-  }
-
-  @Validate(1, 1, "abs", false)
-  static abs(
-    args: Value[],
-    source: string,
-    command: ExprNS.Call,
-    context: Context,
-  ): BigIntValue | NumberValue {
-    const x = args[0];
-    switch (x.type) {
-      case "bigint": {
-        const intVal = x.value;
-        const result: bigint = intVal < 0 ? -intVal : intVal;
-        return { type: "bigint", value: result };
-      }
-      case "number": {
-        return { type: "number", value: Math.abs(x.value) };
-      }
-      case "complex": {
-        // Calculate the modulus (absolute value) of a complex number.
-        const real = x.value.real;
-        const imag = x.value.imag;
-        const modulus = Math.sqrt(real * real + imag * imag);
-        return { type: "number", value: modulus };
-      }
-      default:
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[0].type, "float', 'int' or 'complex"),
-        );
-    }
-  }
-
-  @Validate(1, 1, "len", true)
-  static len(args: Value[], source: string, command: ExprNS.Call, context: Context): BigIntValue {
-    const val = args[0];
-    if (val.type === "string" || val.type === "list") {
-      // The spread operator is used to count the number of Unicode code points
-      // in the string
-      return { type: "bigint", value: BigInt([...val.value].length) };
-    }
-    handleRuntimeError(
-      context,
-      new TypeError(source, command, context, val.type, "object with length"),
-    );
-  }
-
-  static error(args: Value[], _source: string, command: ExprNS.Call, context: Context): Value {
-    const output = "Error: " + args.map(arg => toPythonString(arg)).join(" ") + "\n";
-    handleRuntimeError(context, new UserError(output, command));
-  }
-
+export class MathBuiltins {
   @Validate(1, 1, "math_acos", false)
   static math_acos(
     args: Value[],
@@ -659,7 +349,7 @@ export class BuiltInFunctions {
       );
     }
 
-    const erfc = 1 - BuiltInFunctions.math_erf([args[0]], source, command, context).value;
+    const erfc = 1 - MathBuiltins.math_erf([args[0]], source, command, context).value;
 
     return { type: "number", value: erfc };
   }
@@ -757,7 +447,7 @@ export class BuiltInFunctions {
 
     let currentGcd: bigint = values[0] < 0 ? -values[0] : values[0];
     for (let i = 1; i < values.length; i++) {
-      currentGcd = BuiltInFunctions.gcdOfTwo(currentGcd, values[i] < 0 ? -values[i] : values[i]);
+      currentGcd = MathBuiltins._gcdOfTwo(currentGcd, values[i] < 0 ? -values[i] : values[i]);
       if (currentGcd === BigInt(1)) {
         break;
       }
@@ -766,7 +456,7 @@ export class BuiltInFunctions {
     return { type: "bigint", value: currentGcd };
   }
 
-  static gcdOfTwo(a: bigint, b: bigint): bigint {
+  static _gcdOfTwo(a: bigint, b: bigint): bigint {
     let x: bigint = a;
     let y: bigint = b;
     while (y !== BigInt(0)) {
@@ -836,9 +526,9 @@ export class BuiltInFunctions {
       return { type: "bigint", value: BigInt(0) };
     }
 
-    let currentLcm: bigint = BuiltInFunctions.absBigInt(values[0]);
+    let currentLcm: bigint = MathBuiltins._absBigInt(values[0]);
     for (let i = 1; i < values.length; i++) {
-      currentLcm = BuiltInFunctions.lcmOfTwo(currentLcm, BuiltInFunctions.absBigInt(values[i]));
+      currentLcm = MathBuiltins._lcmOfTwo(currentLcm, MathBuiltins._absBigInt(values[i]));
       if (currentLcm === BigInt(0)) {
         break;
       }
@@ -847,12 +537,12 @@ export class BuiltInFunctions {
     return { type: "bigint", value: currentLcm };
   }
 
-  static lcmOfTwo(a: bigint, b: bigint): bigint {
-    const gcdVal: bigint = BuiltInFunctions.gcdOfTwo(a, b);
+  static _lcmOfTwo(a: bigint, b: bigint): bigint {
+    const gcdVal: bigint = MathBuiltins._gcdOfTwo(a, b);
     return BigInt((a / gcdVal) * b);
   }
 
-  static absBigInt(x: bigint): bigint {
+  static _absBigInt(x: bigint): bigint {
     return x < 0 ? -x : x;
   }
 
@@ -981,7 +671,7 @@ export class BuiltInFunctions {
   }
 
   // Computes the product of a and b along with the rounding error using Dekker's algorithm.
-  static twoProd(a: number, b: number): { prod: number; err: number } {
+  static _twoProd(a: number, b: number): { prod: number; err: number } {
     const prod = a * b;
     const c = 134217729; // 2^27 + 1
     const a_hi = a * c - (a * c - a);
@@ -993,7 +683,7 @@ export class BuiltInFunctions {
   }
 
   // Computes the sum of a and b along with the rounding error using Fast TwoSum.
-  static twoSum(a: number, b: number): { sum: number; err: number } {
+  static _twoSum(a: number, b: number): { sum: number; err: number } {
     const sum = a + b;
     const v = sum - a;
     const err = a - (sum - v) + (b - v);
@@ -1001,14 +691,14 @@ export class BuiltInFunctions {
   }
 
   // Performs a fused multiply-add operation: computes (x * y) + z with a single rounding.
-  static fusedMultiplyAdd(x: number, y: number, z: number): number {
-    const { prod, err: prodErr } = BuiltInFunctions.twoProd(x, y);
-    const { sum, err: sumErr } = BuiltInFunctions.twoSum(prod, z);
+  static _fusedMultiplyAdd(x: number, y: number, z: number): number {
+    const { prod, err: prodErr } = MathBuiltins._twoProd(x, y);
+    const { sum, err: sumErr } = MathBuiltins._twoSum(prod, z);
     const result = sum + (prodErr + sumErr);
     return result;
   }
 
-  static toNumber(val: Value, source: string, command: ExprNS.Call, context: Context): number {
+  static _toNumber(val: Value, source: string, command: ExprNS.Call, context: Context): number {
     if (val.type === "bigint") {
       return Number(val.value);
     } else if (val.type === "number") {
@@ -1028,9 +718,9 @@ export class BuiltInFunctions {
     command: ExprNS.Call,
     context: Context,
   ): NumberValue {
-    const xVal = BuiltInFunctions.toNumber(args[0], source, command, context);
-    const yVal = BuiltInFunctions.toNumber(args[1], source, command, context);
-    const zVal = BuiltInFunctions.toNumber(args[2], source, command, context);
+    const xVal = MathBuiltins._toNumber(args[0], source, command, context);
+    const yVal = MathBuiltins._toNumber(args[1], source, command, context);
+    const zVal = MathBuiltins._toNumber(args[2], source, command, context);
 
     // Special-case handling: According to the IEEE 754 standard, fma(0, inf, nan)
     // and fma(inf, 0, nan) should return NaN.
@@ -1044,15 +734,15 @@ export class BuiltInFunctions {
       return { type: "number", value: NaN };
     }
 
-    const result = BuiltInFunctions.fusedMultiplyAdd(xVal, yVal, zVal);
+    const result = MathBuiltins._fusedMultiplyAdd(xVal, yVal, zVal);
     return { type: "number", value: result };
   }
 
   @Validate(2, 2, "math_fmod", false)
   static math_fmod(args: Value[], source: string, command: ExprNS.Call, context: Context): Value {
     // Convert inputs to numbers
-    const xVal = BuiltInFunctions.toNumber(args[0], source, command, context);
-    const yVal = BuiltInFunctions.toNumber(args[1], source, command, context);
+    const xVal = MathBuiltins._toNumber(args[0], source, command, context);
+    const yVal = MathBuiltins._toNumber(args[1], source, command, context);
 
     // Divisor cannot be zero
     if (yVal === 0) {
@@ -1068,7 +758,7 @@ export class BuiltInFunctions {
     return { type: "number", value: remainder };
   }
 
-  static roundToEven(num: number): number {
+  static _roundToEven(num: number): number {
     //uses Banker's Rounding as per Python's Round() function
     const floorVal = Math.floor(num);
     const ceilVal = Math.ceil(num);
@@ -1122,7 +812,7 @@ export class BuiltInFunctions {
     }
 
     const quotient = xValue / yValue;
-    const n = BuiltInFunctions.roundToEven(quotient);
+    const n = MathBuiltins._roundToEven(quotient);
     const remainder = xValue - n * yValue;
 
     return { type: "number", value: remainder };
@@ -1259,7 +949,7 @@ export class BuiltInFunctions {
     command: ExprNS.Call,
     context: Context,
   ): NumberValue {
-    const xVal = BuiltInFunctions.toNumber(args[0], source, command, context);
+    const xVal = MathBuiltins._toNumber(args[0], source, command, context);
 
     if (args[1].type !== "bigint") {
       handleRuntimeError(context, new TypeError(source, command, context, args[1].type, "int"));
@@ -1419,7 +1109,7 @@ export class BuiltInFunctions {
       );
     }
 
-    const z = BuiltInFunctions.toNumber(x, source, command, context);
+    const z = MathBuiltins._toNumber(x, source, command, context);
     const result = gamma(z);
 
     return { type: "number", value: result };
@@ -1440,7 +1130,7 @@ export class BuiltInFunctions {
       );
     }
 
-    const z = BuiltInFunctions.toNumber(x, source, command, context);
+    const z = MathBuiltins._toNumber(x, source, command, context);
     const result = lgamma(z);
 
     return { type: "number", value: result };
@@ -1764,564 +1454,27 @@ export class BuiltInFunctions {
     const result = Math.sqrt(num);
     return { type: "number", value: result };
   }
-
-  @Validate(2, null, "max", true)
-  static max(args: Value[], source: string, command: ExprNS.Call, context: Context): Value {
-    const numericTypes = ["bigint", "number"];
-    const firstType = args[0].type;
-    const isNumericValue = numericTypes.includes(firstType);
-    const isString = firstType === "string";
-
-    for (let i = 1; i < args.length; i++) {
-      const t = args[i].type;
-      if (isNumericValue && !numericTypes.includes(t)) {
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[i].type, "float' or 'int"),
-        );
-      }
-      if (isString && t !== "string") {
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[i].type, "string"),
-        );
-      }
-    }
-
-    let useFloat = false;
-    if (isNumericValue) {
-      for (const arg of args) {
-        if (arg.type === "number") {
-          useFloat = true;
-          break;
-        }
-      }
-    }
-
-    let maxIndex = 0;
-    if (isNumericValue) {
-      if (useFloat) {
-        if (args[0].type !== "number" && args[0].type !== "bigint") {
-          handleRuntimeError(
-            context,
-            new TypeError(source, command, context, args[0].type, "float' or 'int"),
-          );
-        }
-        let maxVal: number = Number(args[0].value);
-        for (let i = 1; i < args.length; i++) {
-          const arg = args[i];
-          if (!isNumeric(arg)) {
-            handleRuntimeError(
-              context,
-              new TypeError(source, command, context, arg.type, "float' or 'int"),
-            );
-          }
-          const curr: number = Number(arg.value);
-          if (curr > maxVal) {
-            maxVal = curr;
-            maxIndex = i;
-          }
-        }
-      } else {
-        if (args[0].type !== "bigint") {
-          handleRuntimeError(context, new TypeError(source, command, context, args[0].type, "int"));
-        }
-        let maxVal: bigint = args[0].value;
-        for (let i = 1; i < args.length; i++) {
-          const arg = args[i];
-          if (arg.type !== "bigint") {
-            handleRuntimeError(context, new TypeError(source, command, context, arg.type, "int"));
-          }
-          const curr: bigint = arg.value;
-          if (curr > maxVal) {
-            maxVal = curr;
-            maxIndex = i;
-          }
-        }
-      }
-    } else if (isString) {
-      if (args[0].type !== "string") {
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[0].type, "string"),
-        );
-      }
-      let maxVal = args[0].value;
-      for (let i = 1; i < args.length; i++) {
-        const arg = args[i];
-        if (arg.type !== "string") {
-          handleRuntimeError(context, new TypeError(source, command, context, arg.type, "string"));
-        }
-        const curr = arg.value;
-        if (curr > maxVal) {
-          maxVal = curr;
-          maxIndex = i;
-        }
-      }
-    } else {
-      // Won't happen
-      throw new Error(`max: unsupported type ${firstType}`);
-    }
-
-    return args[maxIndex];
-  }
-
-  @Validate(2, null, "min", true)
-  static min(args: Value[], source: string, command: ExprNS.Call, context: Context): Value {
-    if (args.length < 2) {
-      handleRuntimeError(
-        context,
-        new MissingRequiredPositionalError(source, command, "min", Number(2), args, true),
-      );
-    }
-
-    const numericTypes = ["bigint", "number"];
-    const firstType = args[0].type;
-    const isNumericValue = numericTypes.includes(firstType);
-    const isString = firstType === "string";
-
-    for (let i = 1; i < args.length; i++) {
-      const t = args[i].type;
-      if (isNumericValue && !numericTypes.includes(t)) {
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[i].type, "float' or 'int"),
-        );
-      }
-      if (isString && t !== "string") {
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[i].type, "string"),
-        );
-      }
-    }
-
-    let useFloat = false;
-    if (isNumericValue) {
-      for (const arg of args) {
-        if (arg.type === "number") {
-          useFloat = true;
-          break;
-        }
-      }
-    }
-
-    let maxIndex = 0;
-    if (isNumericValue) {
-      if (useFloat) {
-        if (args[0].type !== "number" && args[0].type !== "bigint") {
-          handleRuntimeError(
-            context,
-            new TypeError(source, command, context, args[0].type, "float' or 'int"),
-          );
-        }
-        let maxVal: number = Number(args[0].value);
-        for (let i = 1; i < args.length; i++) {
-          const arg = args[i];
-          if (!isNumeric(arg)) {
-            handleRuntimeError(
-              context,
-              new TypeError(source, command, context, arg.type, "float' or 'int"),
-            );
-          }
-          const curr: number = Number(arg.value);
-          if (curr < maxVal) {
-            maxVal = curr;
-            maxIndex = i;
-          }
-        }
-      } else {
-        if (args[0].type !== "bigint") {
-          handleRuntimeError(context, new TypeError(source, command, context, args[0].type, "int"));
-        }
-        let maxVal: bigint = args[0].value;
-        for (let i = 1; i < args.length; i++) {
-          const arg = args[i];
-          if (arg.type !== "bigint") {
-            handleRuntimeError(context, new TypeError(source, command, context, arg.type, "int"));
-          }
-          const curr: bigint = arg.value;
-          if (curr < maxVal) {
-            maxVal = curr;
-            maxIndex = i;
-          }
-        }
-      }
-    } else if (isString) {
-      if (args[0].type !== "string") {
-        handleRuntimeError(
-          context,
-          new TypeError(source, command, context, args[0].type, "string"),
-        );
-      }
-      let maxVal = args[0].value;
-      for (let i = 1; i < args.length; i++) {
-        const arg = args[i];
-        if (arg.type !== "string") {
-          handleRuntimeError(context, new TypeError(source, command, context, arg.type, "string"));
-        }
-        const curr = arg.value;
-        if (curr < maxVal) {
-          maxVal = curr;
-          maxIndex = i;
-        }
-      }
-    } else {
-      // Won't happen
-      throw new Error(`min: unsupported type ${firstType}`);
-    }
-
-    return args[maxIndex];
-  }
-
-  @Validate(null, 0, "random_random", true)
-  static random_random(
-    _args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): NumberValue {
-    const result = Math.random();
-    return { type: "number", value: result };
-  }
-
-  @Validate(1, 2, "round", true)
-  static round(
-    args: Value[],
-    source: string,
-    command: ExprNS.Call,
-    context: Context,
-  ): NumberValue | BigIntValue {
-    const numArg = args[0];
-    if (!isNumeric(numArg)) {
-      handleRuntimeError(
-        context,
-        new TypeError(source, command, context, numArg.type, "float' or 'int"),
-      );
-    }
-
-    let ndigitsArg: BigIntValue = { type: "bigint", value: BigInt(0) };
-    if (args.length === 2 && args[1].type !== "none") {
-      if (args[1].type !== "bigint") {
-        handleRuntimeError(context, new TypeError(source, command, context, args[1].type, "int"));
-      }
-      ndigitsArg = args[1];
-    } else {
-      const shifted = Intl.NumberFormat("en-US", {
-        roundingMode: "halfEven",
-        useGrouping: false,
-        maximumFractionDigits: 0,
-      } as Intl.NumberFormatOptions).format(numArg.value);
-      return { type: "bigint", value: BigInt(shifted) };
-    }
-
-    if (numArg.type === "number") {
-      const numberValue: number = numArg.value;
-      if (ndigitsArg.value >= 0) {
-        const shifted = Intl.NumberFormat("en-US", {
-          roundingMode: "halfEven",
-          useGrouping: false,
-          maximumFractionDigits: Number(ndigitsArg.value),
-        } as Intl.NumberFormatOptions).format(numberValue);
-        return { type: "number", value: Number(shifted) };
-      } else {
-        const shifted = Intl.NumberFormat("en-US", {
-          roundingMode: "halfEven",
-          useGrouping: false,
-          maximumFractionDigits: 0,
-        } as Intl.NumberFormatOptions).format(numArg.value / 10 ** -Number(ndigitsArg.value));
-        return { type: "number", value: Number(shifted) * 10 ** -Number(ndigitsArg.value) };
-      }
-    } else {
-      if (ndigitsArg.value >= 0) {
-        return numArg;
-      } else {
-        const shifted = Intl.NumberFormat("en-US", {
-          roundingMode: "halfEven",
-          useGrouping: false,
-          maximumFractionDigits: 0,
-        } as Intl.NumberFormatOptions).format(
-          Number(numArg.value) / 10 ** -Number(ndigitsArg.value),
-        );
-        return { type: "bigint", value: BigInt(shifted) * 10n ** -ndigitsArg.value };
-      }
-    }
-  }
-
-  @Validate(null, 0, "time_time", true)
-  static time_time(
-    _args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): NumberValue {
-    const currentTime = Date.now();
-    return { type: "number", value: currentTime };
-  }
-
-  @Validate(1, 1, "is_none", true)
-  static is_none(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return { type: "bool", value: obj.type === "none" };
-  }
-
-  @Validate(1, 1, "is_float", true)
-  static is_float(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return { type: "bool", value: obj.type === "number" };
-  }
-
-  @Validate(1, 1, "is_string", true)
-  static is_string(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return { type: "bool", value: obj.type === "string" };
-  }
-
-  @Validate(1, 1, "is_boolean", true)
-  static is_boolean(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return { type: "bool", value: obj.type === "bool" };
-  }
-
-  @Validate(1, 1, "is_complex", true)
-  static is_complex(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return { type: "bool", value: obj.type === "complex" };
-  }
-
-  @Validate(1, 1, "is_int", true)
-  static is_int(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return { type: "bool", value: obj.type === "bigint" };
-  }
-
-  @Validate(1, 1, "is_function", true)
-  static is_function(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): BoolValue {
-    const obj = args[0];
-    return {
-      type: "bool",
-      value: obj.type === "function" || obj.type === "closure" || obj.type === "builtin",
-    };
-  }
-
-  static async input(
-    _args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    context: Context,
-  ): Promise<Value> {
-    const userInput = await receiveInput(context);
-    return { type: "string", value: userInput };
-  }
-
-  static async print(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    context: Context,
-  ): Promise<Value> {
-    const output = args.map(arg => toPythonString(arg)).join(" ");
-    await displayOutput(context, output);
-    return { type: "none" };
-  }
-  static str(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): StringValue {
-    if (args.length === 0) {
-      return { type: "string", value: "" };
-    }
-    const obj = args[0];
-    const result = toPythonString(obj);
-    return { type: "string", value: result };
-  }
-  @Validate(1, 1, "repr", true)
-  static repr(
-    args: Value[],
-    _source: string,
-    _command: ExprNS.Call,
-    _context: Context,
-  ): StringValue {
-    const obj = args[0];
-    const result = toPythonString(obj, true);
-    return { type: "string", value: result };
-  }
 }
 
-import { ExprNS } from "./ast-types";
-import { isFalsy } from "./engines/cse/operators";
-import { isNumeric } from "./engines/cse/utils";
-import py_s1_constants from "./stdlib/py_s1_constants.json";
-import { PyComplexNumber } from "./types";
-
-// NOTE: If we ever switch to another Python “chapter” (e.g. py_s2_constants),
-//       just change the variable below to switch to the set.
-const constants = py_s1_constants;
-
-/*
-    Create a map to hold built-in constants.
-    Each constant is stored with a string key and its corresponding value object.
-*/
-export const builtInConstants = new Map<string, Value>();
-
-const constantMap = {
-  math_e: { type: "number", value: Math.E },
-  math_inf: { type: "number", value: Infinity },
-  math_nan: { type: "number", value: NaN },
-  math_pi: { type: "number", value: Math.PI },
-  math_tau: { type: "number", value: 2 * Math.PI },
-} as const;
-
-for (const name of constants.constants) {
-  const valueObj = constantMap[name as keyof typeof constantMap];
-  if (!valueObj) {
-    throw new Error(`Constant '${name}' is not implemented`);
+for (const builtin of Object.getOwnPropertyNames(MathBuiltins)) {
+  if (
+    typeof MathBuiltins[builtin as keyof typeof MathBuiltins] === "function" &&
+    !builtin.startsWith("_")
+  ) {
+    mathBuiltins.set(builtin, {
+      type: "builtin",
+      func: MathBuiltins[builtin as keyof typeof MathBuiltins] as BuiltinValue["func"],
+      name: builtin,
+      minArgs: minArgMap.get(builtin) || 0,
+    });
   }
-  builtInConstants.set(name, valueObj);
+}
+for (const [name, info] of Object.entries(constantMap)) {
+  mathBuiltins.set(name, info);
 }
 
-/*
-    Create a map to hold built-in functions.
-    The keys are strings (function names) and the values are functions that can take any arguments.
-*/
-export const builtIns = new Map<string, BuiltinValue>();
-for (const name of constants.builtInFuncs) {
-  const impl = BuiltInFunctions[name as keyof BuiltInFunctions];
-  if (typeof impl !== "function") {
-    throw new Error(`BuiltInFunctions.${name} is not implemented`);
-  }
-  const builtinName = name.startsWith("_") ? name.substring(1) : name;
-  builtIns.set(name, {
-    type: "builtin",
-    name: builtinName,
-    func: impl,
-    minArgs: minArgMap.get(name) || 0,
-  });
-}
-
-/**
- * Converts a number to a string that mimics Python's float formatting behavior.
- *
- * In Python, float values are printed in scientific notation when their absolute value
- * is ≥ 1e16 or < 1e-4. This differs from JavaScript/TypeScript's default behavior,
- * so we explicitly enforce these formatting thresholds.
- *
- * The logic here is based on Python's internal `format_float_short` implementation
- * in CPython's `pystrtod.c`:
- * https://github.com/python/cpython/blob/main/Python/pystrtod.c
- *
- * Special cases such as -0, Infinity, and NaN are also handled to ensure that
- * output matches Python’s display conventions.
- */
-export function toPythonFloat(num: number): string {
-  if (Object.is(num, -0)) {
-    return "-0.0";
-  }
-  if (num === 0) {
-    return "0.0";
-  }
-
-  if (num === Infinity) {
-    return "inf";
-  }
-  if (num === -Infinity) {
-    return "-inf";
-  }
-
-  if (Number.isNaN(num)) {
-    return "nan";
-  }
-
-  if (Math.abs(num) >= 1e16 || (num !== 0 && Math.abs(num) < 1e-4)) {
-    return num.toExponential().replace(/e([+-])(\d)$/, "e$10$2");
-  }
-  if (Number.isInteger(num)) {
-    return num.toFixed(1).toString();
-  }
-  return num.toString();
-}
-function escape(str: string): string {
-  let escaped = JSON.stringify(str);
-  if (!(str.includes("'") && !str.includes('"'))) {
-    escaped = `'${escaped.slice(1, -1).replace(/'/g, "\\'").replace(/\\"/g, '"')}'`;
-  }
-  return escaped;
-}
-function toPythonList(obj: Value): string {
-  return stringify(obj);
-}
-
-export function toPythonString(obj: Value, repr: boolean = false): string {
-  let ret: string = "";
-  if (obj.type == "builtin") {
-    return `<built-in function ${obj.name}>`;
-  }
-  if (obj.type === "bigint" || obj.type === "complex") {
-    ret = obj.value.toString();
-  } else if (obj.type === "number") {
-    ret = toPythonFloat(obj.value);
-  } else if (obj.type === "bool") {
-    if (obj.value) {
-      return "True";
-    } else {
-      return "False";
-    }
-  } else if (obj.type === "error") {
-    return obj.message;
-  } else if (obj.type === "closure") {
-    if (obj.closure.node) {
-      const funcName =
-        obj.closure.node.kind === "FunctionDef" ? obj.closure.node.name.lexeme : "(anonymous)";
-      return `<function ${funcName}>`;
-    }
-  } else if (obj.type === "none") {
-    ret = "None";
-  } else if (obj.type === "string") {
-    ret = repr ? escape(obj.value) : obj.value;
-  } else if (obj.type === "function") {
-    const funcName = obj.name || "(anonymous)";
-    ret = `<function ${funcName}>`;
-  } else if (obj.type === "list") {
-    ret = toPythonList(obj);
-  } else {
-    ret = `<${obj.type} object>`;
-  }
-  return ret;
-}
+export default {
+  name: GroupName.MATH,
+  prelude: "",
+  builtins: mathBuiltins,
+};
