--- src/stdlib/py_s1_constants.json
@@ -1,80 +1,0 @@
-{
-  "builtInFuncs": [
-    "abs",
-    "arity",
-    "bool",
-    "complex",
-    "error",
-    "float",
-    "imag",
-    "int",
-    "math_acos",
-    "math_acosh",
-    "math_asin",
-    "math_asinh",
-    "math_atan",
-    "math_atan2",
-    "math_atanh",
-    "math_cos",
-    "math_cosh",
-    "math_degrees",
-    "math_erf",
-    "math_erfc",
-    "math_comb",
-    "math_factorial",
-    "math_gcd",
-    "math_isqrt",
-    "math_lcm",
-    "math_perm",
-    "math_ceil",
-    "math_fabs",
-    "math_floor",
-    "math_fma",
-    "math_fmod",
-    "math_remainder",
-    "math_trunc",
-    "math_copysign",
-    "math_isfinite",
-    "math_isinf",
-    "math_isnan",
-    "math_ldexp",
-    "math_nextafter",
-    "math_ulp",
-    "math_cbrt",
-    "math_exp",
-    "math_exp2",
-    "math_expm1",
-    "math_gamma",
-    "math_lgamma",
-    "math_log",
-    "math_log10",
-    "math_log1p",
-    "math_log2",
-    "math_pow",
-    "math_radians",
-    "math_sin",
-    "math_sinh",
-    "math_tan",
-    "math_tanh",
-    "math_sqrt",
-    "is_none",
-    "is_boolean",
-    "is_complex",
-    "is_string",
-    "is_function",
-    "is_float",
-    "is_int",
-    "len",
-    "max",
-    "min",
-    "random_random",
-    "real",
-    "round",
-    "time_time",
-    "str",
-    "repr",
-    "print",
-    "input"
-  ],
-  "constants": ["math_e", "math_inf", "math_nan", "math_pi", "math_tau"]
-}
